L4 part I;

+++ c++/L05 while/05_4.cpp
@@ -28,7 +28,7 @@ int main()
 
 int my_sqrt(int n)
 {
-    double i = 1;
+    int i = 1;
     while(i * i < n)
     {
         i++;
